Parse SWISSPROT as name or acc based on a meta flag
In mutation data files, the SWISSPROT column may hold either
UniProtKB/Swiss-Prot names or UniProtKB accessions depending on the
annotation used. This patch makes the loader read it as a name by
default, or as an accession if specified by the optional flag
`swissprot_identifier` in the metadata file.

In both cases, the database column `oncotator_uniprot_accession` will be
filled, which the frontend uses as a key to look up Pfam domains. The
column `oncotator_uniprot_id` in the same table is redundant and now
deprecated.