relaxed the validation for some sample level attributes
this was done for a few attributes
(TUMOR_SITE", "TUMOR_TISSUE_SITE","PRIMARY_SITE", "METASTATIC_SITE",GLEASON_SCORE)
for which the impact on the UI is expected to be minimal. E.g.
METASTATIC_SITE will still appear in clinical data if set in patient
clinical file. The only effect expected is it being omitted from a
header label in the patient/sample view (according to some logic found
in ClinicalAttributesUtil.js)