Expand genes five at a time in order, with issues
Make the Oncoprint model aware of to-be-expanded genes and their
initialisation, so that the setup procedure can register callbacks
to be run via buttons in the track options view. This still has two
main issues: new expansion tracks are always sorted immediately
below the gene set track instead of underneath its pre-existing
expansion tracks, and the list of genes to expand can not be reset
(when all tracks are hidden).